+++ app/src/main/java/io/github/muntashirakon/AppManager/backup/struct/DeleteOpOptions.java
@@ -12,8 +12,6 @@ import androidx.annotation.Nullable;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.util.Objects;
-
 import io.github.muntashirakon.AppManager.history.IJsonSerializer;
 import io.github.muntashirakon.AppManager.utils.JSONUtils;
 
@@ -26,15 +24,15 @@ public class DeleteOpOptions implements Parcelable, IJsonSerializer {
     public final String[] relativeDirs;
 
     public DeleteOpOptions(@NonNull String packageName, @UserIdInt int userId, @Nullable String[] relativeDirs) {
-        this.packageName = packageName;
-        this.userId = userId;
-        this.relativeDirs = relativeDirs;
+        this.packageName = BackupOperationOptionValidator.requirePackageName(packageName);
+        this.userId = BackupOperationOptionValidator.requireUserId(userId);
+        this.relativeDirs = BackupOperationOptionValidator.requireRelativeDirs(relativeDirs);
     }
 
     protected DeleteOpOptions(@NonNull Parcel in) {
-        packageName = Objects.requireNonNull(in.readString());
-        userId = in.readInt();
-        relativeDirs = Objects.requireNonNull(in.createStringArray());
+        packageName = BackupOperationOptionValidator.requirePackageName(in.readString());
+        userId = BackupOperationOptionValidator.requireUserId(in.readInt());
+        relativeDirs = BackupOperationOptionValidator.requireRelativeDirs(in.createStringArray());
     }
 
     public static final Creator<DeleteOpOptions> CREATOR = new Creator<DeleteOpOptions>() {
@@ -58,24 +56,31 @@ public class DeleteOpOptions implements Parcelable, IJsonSerializer {
 
     @Override
     public void writeToParcel(@NonNull Parcel dest, int flags) {
-        dest.writeString(packageName);
-        dest.writeInt(userId);
-        dest.writeStringArray(relativeDirs);
+        dest.writeString(BackupOperationOptionValidator.requirePackageName(packageName));
+        dest.writeInt(BackupOperationOptionValidator.requireUserId(userId));
+        dest.writeStringArray(BackupOperationOptionValidator.requireRelativeDirs(relativeDirs));
     }
 
     public DeleteOpOptions(@NonNull JSONObject jsonObject) throws JSONException {
-        packageName = jsonObject.getString("package_name");
-        userId = jsonObject.getInt("user_id");
-        relativeDirs = JSONUtils.getArray(String.class, jsonObject.optJSONArray("relative_dirs"));
+        try {
+            packageName = BackupOperationOptionValidator.requirePackageName(jsonObject.getString("package_name"));
+            userId = BackupOperationOptionValidator.requireUserId(jsonObject.getInt("user_id"));
+            relativeDirs = BackupOperationOptionValidator.requireRelativeDirs(
+                    BackupOperationOptionValidator.readStringArray(jsonObject, "relative_dirs",
+                            false, "relative directory"));
+        } catch (IllegalArgumentException e) {
+            throw new JSONException(e.getMessage());
+        }
     }
 
     @NonNull
     @Override
     public JSONObject serializeToJson() throws JSONException {
         JSONObject jsonObject = new JSONObject();
-        jsonObject.put("package_name", packageName);
-        jsonObject.put("user_id", userId);
-        jsonObject.put("relative_dirs", JSONUtils.getJSONArray(relativeDirs));
+        jsonObject.put("package_name", BackupOperationOptionValidator.requirePackageName(packageName));
+        jsonObject.put("user_id", BackupOperationOptionValidator.requireUserId(userId));
+        jsonObject.put("relative_dirs", JSONUtils.getJSONArray(
+                BackupOperationOptionValidator.requireRelativeDirs(relativeDirs)));
         return jsonObject;
     }
 }